flask/requirements.txt
-i https://pypi.org/simple
astroid==2.4.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
eventlet==0.25.2
faker==4.1.0
flask-socketio==4.3.0
flask==1.1.2
geographiclib==1.50
geopy==1.22.0
greenlet==0.4.15
isort==4.3.21
itsdangerous==1.1.0
jinja2==2.11.2
lazy-object-proxy==1.4.3
markupsafe==1.1.1
mccabe==0.6.1
mongoengine==0.16.2
monotonic==1.5
pillow==7.1.2
pycparser==2.20
pylint==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
svgwrite==1.4
text-unidecode==1.3
toml==0.10.1
tree==0.2.4
uWSGI==2.0.17.1
werkzeug==1.0.1
wrapt==1.12.1
tqdm